refactor: Move createHighlighter functions to own file
