+++ packages/test-case-component/src/createHighlighter.ts
@@ -1,0 +1,15 @@
+import { createHighlighter as shikiCreateHighlighter, createCssVariablesTheme } from "shiki";
+
+const myTheme = createCssVariablesTheme({
+    name: "css-variables",
+    variablePrefix: "--shiki-",
+    variableDefaults: {},
+    fontStyle: true,
+});
+
+export function createHighlighter() {
+    return shikiCreateHighlighter({
+        themes: [myTheme],
+        langs: ["javascript", "typescript", "python", "markdown"],
+    });
+}
